Move reponse.d.ts
