+++ sticker-source/response.d.ts
